修改scale 位置

+++ utils/image_utils.py
@@ -1,10 +1,11 @@
+import numpy as np
 import torch
 
 
 def convert_to_numpy(images):
     # 如果图片不是numpy格式，就转换成numpy格式
     if isinstance(images, torch.Tensor):
-        images = images[:, :, :, [2, 1, 0]].mul(255).detach().squeeze().cpu().numpy()
+        images = images[:, :, :, [2, 1, 0]].mul(255).detach().squeeze().cpu().numpy().astype(np.uint8)
     return images
